detoks: Add dependencies on `build.gradle`
Add a number of dependencies that we will use later in our lightweight
implementation of the `detoks` application. This lightweight
implementation of `detoks` will not run on Android emulators. Instead,
each peer will run as a light process in a local machine.

Through the lightweight implementation, we will be able to test our
transactions engine (which we will introduce in a following commit)
easily and quickly.