detoks: Add `TransactionEngine`
Add `TransactionEngine` which will serve as the main API class for
transactions.

The already existing `DeToksCommunity` will now inherit all of
`TransactionEngine`'s member functions related to transactions.

The `TransactionEngine` also serves as a lightweight implementation
of the `DeToksCommunity`, so that we can use it in our lightweight
version of our Detoks app.