Benchmark to other_group_progress to get UI back (#9)
* detoks: Add dependencies on `build.gradle`

Add a number of dependencies that we will use later in our lightweight
implementation of the `detoks` application. This lightweight
implementation of `detoks` will not run on Android emulators. Instead,
each peer will run as a light process in a local machine.

Through the lightweight implementation, we will be able to test our
transactions engine (which we will introduce in a following commit)
easily and quickly.

* detoks: Add `TransactionEngine`

Add `TransactionEngine` which will serve as the main API class for
transactions.

The already existing `DeToksCommunity` will now inherit all of
`TransactionEngine`'s member functions related to transactions.

The `TransactionEngine` also serves as a lightweight implementation
of the `DeToksCommunity`, so that we can use it in our lightweight
version of our Detoks app.

* detoks: Introduce `IntroduceLightweightTxApp`

Introduce a lightweight implementation of the Detoks application that
can be used for testing and benchmarking.

* Progress with creating functions for benchmarking. So far there is basic block not encrypted with same content, basic block not encrypted with random content, basic block random signed, and unencrypted signed trustchain with in memory storage.

* More functions that can benchmark things.

* added listeners to the transaction api, and functions to send blocks over ipv8

* added functions to receive blocks over ipv8

* refactor into benchmark class

* remove interface again

---------

Co-authored-by: Kostas Andriopoulos <konstantinosand@gmail.com>
Co-authored-by: mehulbhuradia <mehulbhuradia007@gmail.com>
Co-authored-by: tudatt <dosar.andrei.titu@gmail.com>