detoks: Add `BasicBlock` class for benchmarking
Add the `BasicBlock` class that we will use later on to benchmark the
creation of large numbers of simple blocks.

For context, one of our deliverables entails block creation without
using IPv8 at all. For this purpose, we created the `BasicBlock`
class that contains all the necessary information of a basic block.